Alterações Request perguntas ChatGPT

+++ zup/zup/zup/src/main/java/school/sptech/zup/repository/NoticiaRepository.java
@@ -3,5 +3,5 @@ package school.sptech.zup.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import school.sptech.zup.domain.Noticia;
 
-public interface NoticiaRepository extends JpaRepository<Noticia, Long> {
+public interface NoticiaRepository extends JpaRepository<Noticia, Integer> {
 }
